Fix parallel UB and Memory

+++ fem/lininteg.hpp
@@ -673,7 +673,7 @@ public:
       int myid;
       MPI_Comm_rank(comm, &myid);
 
-      int seed = (seed_ > 0) ? seed_ + myid : (int)time(0) + myid;
+      int seed = static_cast<int>((seed_ > 0) ? seed_ + myid : (int)time(0) + myid);
       SetSeed(seed);
    }
 #else
